Fixing signal and slot connection.
svn path=/trunk/koffice/; revision=1001381

+++ krita/plugins/tools/defaulttools/kis_tool_brush.h
@@ -55,7 +55,7 @@ protected:
 private slots:
 
     void timeoutPaint();
-    void slotSetRate(int rate);
+    void slotSetRate(qreal rate);
     void slotSetSmoothness(int smoothness);
     void slotSetMagnetism(int magnetism);
